requirements.txt
blinker==1.9.0
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
Flask==3.1.1
flask-cors==6.0.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
lazy_loader==0.4
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.13.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tifffile==2025.5.10
urllib3==2.4.0
Werkzeug==3.1.3
